+++ activerecord/lib/active_record/connection_adapters/abstract_adapter.rb
@@ -387,12 +387,6 @@ module ActiveRecord
         @transaction.number
       end
 
-      def transaction_joinable=(joinable)
-        message = "#transaction_joinable= is deprecated. Please pass the :joinable option to #begin_transaction instead."
-        ActiveSupport::Deprecation.warn message
-        @transaction.joinable = joinable
-      end
-
       def create_savepoint
       end